+++ adapter/gemini/gemini.go
@@ -61,7 +61,7 @@ func (a *Adapter) LLMInvocationParams(
 	convo *frames.LLMContext, opts adapter.Options,
 ) (Params, error) {
 	fromContext, msgs := a.ExtractInitialSystem(
-		a.SystemWithBuiltins(convo.System()), opts.SystemInstruction,
+		convo.System(), opts.SystemInstruction,
 		convo.MessagesFor(a.IDForLLMSpecificMessages()),
 	)
 	system := a.ResolveSystemInstruction(fromContext, opts.SystemInstruction, true)